Merge `dev`  (#1276)
This PR merges in `dev` so `its-happening` is finally up-to-date and all
PRs against it unblocked. There's really not that much to focus on
except for the consensus update notification being broadcasted by the
subscriber. I only do that if we're "synced". I'm considering to clean
that up a bit though, now it's very much make-it-work.

I do think the `subscriber` is the one that should be broadcasting that
event. I think it's rather annoying though that the definition of
"synced" is now a bit scattered.

---------

Co-authored-by: Chris Schinnerl <chris@sia.tech>