Forgive failed refresh only when contract is in set (#1280)
I noticed on Arequipa that sometimes after a reboot 6 or so contracts
are added to the set which before failed to refresh. That is a result of
the refresh failures not being persisted.

That reminded me that we probably don't want to forgive a failed refresh
if a contract wasn't in the set to begin with since it will only
increase churn if it causes another contract to be pushed out of the set
for the next 6 hours instead.