Allow pinning price settings to an underlying currency (#1294)
This PR adds a `pin` manager to the bus that undertakes action depending
on the user's `pricepinning` settings. Those include pins for gouging
settings as well as autopilot settings. When a setting is pinned, it's
pinned to a fixed value in a certain underlying currency defined in the
settings. `renterd` will periodically fetch the exchange rate for `SC`
against this underlying currency and update the prices accordingly, once
the exchange rate fluctuates over a configured threshold.

`NOTE` I'm sure there's some parts that aren't 100% correct now, I
figured maybe we can build the UI side of the feature so we can test it
using the UI rather than doing it manually. I've never configured pinned
prices on a host so I'm not quite sure how that happens. Do we change
the way we express the various cost settings? E.g. `MaxContractPrice`
for instance, that's 5SC but we probably want that to change along with
price fluctuations but we can't really express 5SC in USD.

---------

Co-authored-by: Chris Schinnerl <chris@sia.tech>