Add support for rewriting only specifc host bases (#1328)
On `gompa` rclone wouldn't work via `localhost` anymore since
`HostBucketEnabled` was set to `true`.
Problem is without this setting, access per proper domain (e.g.
`bucket.example.com`) doesn't work anymore but with it IPs and
`localhost` will be rewritten incorrectly.

The solution is to allow for configuring `HostBucketBases` which is
already supported by `gofakes3` and allows for configuring only specific
domains for which the bucket is extracted from the `host`. That way
`localhost` and `127.0.0.1` still work but if specified
`bucket.example.com` will also work.